1)issued fixed now image will fetch from the database based on the id 2)added validation for the editImage function will allow the only  owner of the image to edit the image otherwise it will redirect to the image list 3)added validation for the deleteImage function will allow the only  owner of the image to delete the image otherwise it will redirect to the image list
1)issued fixed now image will fetch from the database based on the id
2)added validation for the editImage function will allow the only  owner of the image to edit the image otherwise it will redirect to the image list
3)added validation for the deleteImage function will allow the only  owner of the image to delete the image otherwise it will redirect to the image list

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -91,13 +91,20 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    //editImage function is a validation will allow only  owner of the image to edit the image otherwise it will redirect to the image list.
+    //@params:- ImageId,model,session
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
-
+        String error = "Only the owner of the image can edit the image";
         String tags = convertTagsToString(image.getTags());
+        User user = (User) session.getAttribute("loggeduser");
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
+        if(user.getId() != image.getUser().getId()) {
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
         return "images/edit";
     }
 
@@ -139,8 +146,20 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    // deleteImageSubmit function is a validation will allow only  owner of the image to delete the image otherwise it will redirect to the image list.
+    //@params:- ImageId,model,session
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model,HttpSession session) {
+        String error = "Only the owner of the image can delete the image";
+        Image image = imageService.getImage(imageId);
+        String tags = convertTagsToString(image.getTags());
+        User user = (User) session.getAttribute("loggeduser");
+        model.addAttribute("image", image);
+        model.addAttribute("tags", tags);
+        if(user.getId() != image.getUser().getId()) {
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
